@@ -1,16 +1,30 @@
-import React, { Component } from 'react';
+import React, { useState } from 'react';
 import './App.css';
+import Person from './Person/Person';
 
-class App extends Component {
-  render() {
-    return (
-      <div className="App">
-        <h1>
-          Hi, I'm a React App
-        </h1>
-      </div>
-    );
-  }
-}
+const app = (props) => {
+	const [ personsState, setPersonsState ] = useState({
+		persons: [ { name: 'Max', age: 28 }, { name: 'Manu', age: 20 }, { name: 'Stephanie', age: 26 } ]
+	});
 
-export default App;
+	const switchNameHander = () => {
+		setPersonsState({
+			persons: [ { name: 'Maximilian', age: 28 }, { name: 'Manu', age: 20 }, { name: 'Stephanie', age: 26 } ]
+		});
+	};
+
+	return (
+		<div className='App'>
+			<h1>Hi, I'm a React App</h1>
+			<p>This is really working!</p>
+			<button onClick={switchNameHander}>Switch Name</button>
+			<Person name={personsState.persons[0].name} age={personsState.persons[0].age} />
+			<Person name={personsState.persons[1].name} age={personsState.persons[1].age}>
+				My Hobbies: Racing
+			</Person>
+			<Person name={personsState.persons[2].name} age={personsState.persons[2].age} />
+		</div>
+	);
+};
+
+export default app;
